Add files via upload
Pneumonia, or lung infection, is a viral disease that affects lung tissue and interferes with the
normal oxygen exchange between air and blood. Inflammatory secretions that enter the air sacs of
the lungs do not allow the body to properly supply oxygen. And when the disease grabs most of the
lungs, acute respiratory failure develops. Pneumonia affects people with weak immunity, children
and the elderly. In Germany alone, 1.5 million people get the disease each year, 30% of whom are
infants and people over the age of 70. But pneumonia can be treated and needs to be treated! And
it’s a good idea to do this with an integrated approach. Treatment of the disease begins with a
correct diagnosis. Therefore, seek expert advice at the first signs of pathological development.Learn
what the first symptoms of pneumonia are, how to properly and comprehensively treat them, and
what you need to do to diagnose the disease. ## What is the aim of this notebook? I was hired
by a ”Medtronic” as a Data Scientist. Medtronic is a company that produces medical equipment
that ease the life of doctors by digital solutions. In July we start to develop a x-ray machine that
will be able to detect pneumonia. Now we are trying to develop a machine that makes the x-ray
and gives a diagnosis at the same time. By doing so we are planning to eliminate a human factor.